Image refactor fast follows (#53)
* Add 24hour cache control to newly uploaded images

* Make small tweaks to print modal button layout

* Refactor update routes for code cleanliness

* Recalculate optimum width on every hover

* Pass in object position to cart items

* Change cart items layout to use object position

* Add meta description and serve build with cachecontrol set to 1 hour

* Include data storage helpers and use in artwork delete route

* Remove data-text attribute from mobile button style

* Add brandbackdrop to social media links

* Remove commented code and revert animation

+++ routes/delete.js
@@ -1,5 +1,6 @@
 const express = require("express")
 const router = express.Router()
+const { hasSharedResource } = require("../data/helpers/mongodb")
 
 let auth
 if(process.env.NODE_ENV !== 'production') {
@@ -20,7 +21,8 @@ router.post("/:destination", async (req, res) => {
     const model = require(`../models/${destination}`)
 
     try {
-        await storage.bucket(bucket).file(img).delete()
+        const isShared = await hasSharedResource(img, destination)
+        if (!isShared) await storage.bucket(bucket).file(img).delete()
         await model.deleteOne({_id: _id})
         res.json({msg: `${title} was deleted`})
     } catch (error) {
